+++ OgreRenderingModule/OgreMeshAsset.cpp
@@ -59,7 +59,7 @@ bool OgreMeshAsset::DeserializeFromData(const u8 *data_, size_t numBytes)
     // Generate extremity points to submeshes, 1 should be enough
     try
     {
-        for(uint i = 0; i < ogreMesh->getNumSubMeshes(); ++i)
+        for(unsigned short i = 0; i < ogreMesh->getNumSubMeshes(); ++i)
         {
             Ogre::SubMesh *smesh = ogreMesh->getSubMesh(i);
             if (smesh)
@@ -113,7 +113,7 @@ void OgreMeshAsset::SetDefaultMaterial()
         return;
 
 //    originalMaterials.clear();
-    for(uint i = 0; i < ogreMesh->getNumSubMeshes(); ++i)
+    for(unsigned short i = 0; i < ogreMesh->getNumSubMeshes(); ++i)
     {
         Ogre::SubMesh *submesh = ogreMesh->getSubMesh(i);
         if (submesh)
